+++ elephant/spectral.py
@@ -3,7 +3,7 @@
 Identification of spectral properties in analog signals (e.g., the power
 spectrum).
 
-:copyright: Copyright 2015-2016 by the Elephant team, see AUTHORS.txt.
+:copyright: Copyright 2015-2016 by the Elephant team, see `doc/authors.rst`.
 :license: Modified BSD, see LICENSE.txt for details.
 """