+++ elephant/test/test_asset.py
@@ -2,7 +2,7 @@
 """
 Unit tests for the ASSET analysis.
 
-:copyright: Copyright 2014-2016 by the Elephant team, see AUTHORS.txt.
+:copyright: Copyright 2014-2016 by the Elephant team, see `doc/authors.rst`.
 :license: Modified BSD, see LICENSE.txt for details.
 """