+++ elephant/test/test_signal_processing.py
@@ -2,7 +2,7 @@
 """
 Unit tests for the signal_processing module.
 
-:copyright: Copyright 2014-2016 by the Elephant team, see AUTHORS.txt.
+:copyright: Copyright 2014-2016 by the Elephant team, see `doc/authors.rst`.
 :license: Modified BSD, see LICENSE.txt for details.
 """
 from __future__ import division, print_function
@@ -380,13 +380,18 @@ class ButterTestCase(unittest.TestCase):
         self.assertAlmostEqual(psd[0, 256], 0)
 
     def test_butter_filter_function(self):
+        """
+        `elephant.signal_processing.butter` return values test for all
+        available filters (result has to be almost equal):
+            * lfilter
+            * filtfilt
+            * sosfiltfilt
+        """
         # generate white noise AnalogSignal
         noise = neo.AnalogSignal(
             np.random.normal(size=5000),
             sampling_rate=1000 * pq.Hz, units='mV')
 
-        # test if the filter performance is as well with filftunc=lfilter as
-        # with filtfunc=filtfilt (i.e. default option)
         kwds = {'signal': noise, 'highpass_freq': 250.0 * pq.Hz,
                 'lowpass_freq': None, 'filter_function': 'filtfilt'}
         filtered_noise = elephant.signal_processing.butter(**kwds)
@@ -398,7 +403,13 @@ class ButterTestCase(unittest.TestCase):
         _, psd_lfilter = spsig.welch(
             filtered_noise.T, nperseg=1024, fs=1000.0, detrend=lambda x: x)
 
+        kwds['filter_function'] = 'sosfiltfilt'
+        filtered_noise = elephant.signal_processing.butter(**kwds)
+        _, psd_sosfiltfilt = spsig.welch(
+            filtered_noise.T, nperseg=1024, fs=1000.0, detrend=lambda x: x)
+
         self.assertAlmostEqual(psd_filtfilt[0, 0], psd_lfilter[0, 0])
+        self.assertAlmostEqual(psd_filtfilt[0, 0], psd_sosfiltfilt[0, 0])
 
     def test_butter_invalid_filter_function(self):
         # generate a dummy AnalogSignal